Split field and register, haven't implemented encode or tests yet.
